+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## Release candidate
+## v23.39-beta
 
 ### Fix
 - Fix private key provisioning (#267, @vosmol)
@@ -11,6 +11,9 @@
 - Support tenant search in old MongoDB versions (#268, PLUM Sprint 230908)
 - Public invitations to tenant (#261, PLUM Sprint 230908)
 - Human-readable tenant label (#285, PLUM Sprint 230908)
+- Log failed password change requests (#291, PLUM Sprint 230908)
+- Get Github user email address (#289, PLUM Sprint 230908)
+- External login ID token validation (#292, PLUM Sprint 230908)
 
 ---